add signing logic to modules/signer
