+++ rust/tw_bitcoin/src/modules/mod.rs
@@ -1,5 +1,7 @@
 mod input_builder;
 mod output_builder;
+mod signer;
 
+pub use signer::Signer;
 pub use input_builder::InputBuilder;
 pub use output_builder::OutputBuilder;
